TransactionDetail: use HttpClient directly
The background_downloader package was unfortunately causing issues with
the NotificationListener service, since it somehow was unable to
register it's (android internal) callback functions, and so it wasn't able
to report progress properly. Now we're using the dart:io internal
HttpClient with streams. This is enough for our usecase since we don't
really do it "in the background" anyways, and shouldn't be handling
very large files.

Plus, one less dependency which could break or be outdated :)