try new way to handle the app after notification
(ref #92)

This does not solve the issue. It will put the app to the background and
(basically at the same time) set the navigator to the default NavPage
when the app has no previous stack (i.e. was launched from notification)
This should help when you resume from the closed app, but is not really a
"nice" way :( It will open the NavPage in the background, causing
"useless" network traffic.